Change the `BookStore` type from an `interface` to a concrete `class`.
There is no need for the `BookStore` type to be an interface. This is the business logic, so other subsystems should plug in to this code, not the other way around. This is the code that needs to be protected from changing implementation details. The frontend *is* an implementation detail and does not need such protection.